refactor(`host/home`): apply code review — move logos to src/assets, co-locate data, rename components
- Move SVG logos from public/logos/ to src/assets/logos/ and import them as ES modules
- Add src/declarations.d.ts for *.svg and *.css module declarations
- Co-locate APP_SECTIONS data in HomeView.tsx; delete config/apps.ts
- Rename CARD_BASE, HOVER_COLOR_MAP constants to SCREAMING_SNAKE_CASE
- Rename AppIcon → CardIcon; inline AppBadge as <span> in card content
- Use React.ComponentPropsWithoutRef<'div'> on CardIcon
- Rename AppSection.tsx → AppSectionView.tsx; move AppSection/AppCard types to component files
- Add featured flag on AppSection so AppSectionView reuses one component for both layouts
