+++ apps/host/src/components/AppCard.tsx
@@ -1,12 +1,23 @@
+import React from 'react';
 import { Link } from 'react-router';
 
-import type { AppCard, AppColor } from '~/config/apps';
+export type AppColor = 'pink' | 'blue' | 'orange' | 'green' | 'purple';
+
+export interface AppCard {
+  id: string;
+  title: string;
+  description: string;
+  icon: string;
+  color: AppColor;
+  href: string;
+  badge?: string;
+}
 import { cn } from '~/helpers/cn';
 
-const cardBase =
+const CARD_BASE =
   'tw:group tw:flex tw:rounded-[14px] tw:border tw:bg-background tw:p-4 tw:transition-colors tw:hover:bg-muted/50';
 
-const hoverColorMap: Record<AppColor, string> = {
+const HOVER_COLOR_MAP: Record<AppColor, string> = {
   pink: 'tw:group-hover:text-pink-500',
   blue: 'tw:group-hover:text-blue-600',
   orange: 'tw:group-hover:text-orange-500',
@@ -14,14 +25,14 @@ const hoverColorMap: Record<AppColor, string> = {
   purple: 'tw:group-hover:text-purple-500',
 };
 
-function AppIcon({
+function CardIcon({
   icon,
   color,
   className,
-}: {
+  ...props
+}: React.ComponentPropsWithoutRef<'div'> & {
   icon: string;
   color: AppColor;
-  className?: string;
 }) {
   return (
     <div
@@ -29,23 +40,16 @@ function AppIcon({
         'tw:relative tw:flex tw:size-16 tw:shrink-0 tw:items-center tw:justify-center tw:overflow-hidden tw:rounded-[14px] tw:border tw:bg-white tw:p-2',
         className,
       )}
+      {...props}
     >
       <img src={icon} alt="" className="tw:h-full tw:w-full tw:object-contain" />
       <div
         className={cn(
           'tw:pointer-events-none tw:absolute tw:inset-0 tw:bg-[#0064ff] tw:mix-blend-color tw:transition-opacity tw:duration-200 tw:will-change-[opacity] tw:group-hover:opacity-0',
-          hoverColorMap[color],
+          HOVER_COLOR_MAP[color],
         )}
       />
     </div>
-  );
-}
-
-function AppBadge({ label }: { label: string }) {
-  return (
-    <span className="tw:rounded-full tw:bg-blue-100 tw:px-2 tw:py-0.5 tw:text-xs tw:font-medium tw:text-blue-700">
-      {label}
-    </span>
   );
 }
 
@@ -54,11 +58,15 @@ type AppCardProps = Pick<AppCard, 'title' | 'description' | 'icon' | 'color' | '
 function CardContent({ title, description, icon, color, badge }: Omit<AppCardProps, 'href'>) {
   return (
     <>
-      <AppIcon icon={icon} color={color} />
+      <CardIcon icon={icon} color={color} />
       <div className="tw:flex tw:flex-col tw:gap-2">
         <div className="tw:flex tw:items-center tw:gap-2">
           <h3 className="tw:font-semibold tw:text-foreground">{title}</h3>
-          {badge && <AppBadge label={badge} />}
+          {badge && (
+            <span className="tw:rounded-full tw:bg-blue-100 tw:px-2 tw:py-0.5 tw:text-xs tw:font-medium tw:text-blue-700">
+              {badge}
+            </span>
+          )}
         </div>
         <p className="tw:line-clamp-3 tw:text-sm tw:text-muted-foreground">{description}</p>
       </div>
@@ -67,7 +75,7 @@ function CardContent({ title, description, icon, color, badge }: Omit<AppCardPro
 }
 
 export function AppCard({ title, description, icon, color, href, badge }: AppCardProps) {
-  const className = cn(cardBase, 'tw:flex-col tw:gap-4');
+  const className = cn(CARD_BASE, 'tw:flex-col tw:gap-4');
 
   if (href.startsWith('http')) {
     return (
@@ -98,17 +106,21 @@ export function AppCard({ title, description, icon, color, href, badge }: AppCar
 
 export function FeaturedAppCard({ title, description, icon, color, href, badge }: AppCardProps) {
   const className = cn(
-    cardBase,
+    CARD_BASE,
     'tw:h-[132px] tw:flex-row tw:items-center tw:gap-4 tw:border-[#C8C8C8] tw:bg-white',
   );
 
   const content = (
     <>
-      <AppIcon icon={icon} color={color} />
+      <CardIcon icon={icon} color={color} />
       <div className="tw:flex tw:flex-1 tw:flex-col tw:gap-2">
         <div className="tw:flex tw:items-center tw:gap-2">
           <h3 className="tw:font-semibold tw:text-foreground">{title}</h3>
-          {badge && <AppBadge label={badge} />}
+          {badge && (
+            <span className="tw:rounded-full tw:bg-blue-100 tw:px-2 tw:py-0.5 tw:text-xs tw:font-medium tw:text-blue-700">
+              {badge}
+            </span>
+          )}
         </div>
         <p className="tw:text-sm tw:text-muted-foreground">{description}</p>
       </div>
