+++ apps/host/src/declarations.d.ts
@@ -1,0 +1,6 @@
+declare module '*.svg' {
+  const src: string;
+  export default src;
+}
+
+declare module '*.css';
